+++ R/caching.R
@@ -184,12 +184,13 @@ get_hashed_repo_to_dir_mapping <- function(local_repos) {
 #' @return A data frame, one row per checked out repository with columns
 #' repo, host and cache_dir
 rec_checkout_internal_deps <- function(repos_to_process, ref,
-                                      direction = c("upstream"),
+                                      direction = "upstream",
                                       local_repos = get_local_pkgs_from_config(),
                                       verbose = 0) {
   stopifnot(
     is.list(repos_to_process)
   )
+  direction <- check_direction_arg_deprecated(direction)
   check_direction_arg(direction)
   check_verbose_arg(verbose)